refactor: refactor using Ramda js
