+++ basic/arrays/basic-array.js
@@ -1,10 +1,11 @@
+const R = require('Ramda');
 class BasicArray{
+
   sum(numbers){
     if(numbers.length==0){
       throw new Error("Empty Array!");
     }
-    return numbers.reduce((accum, currentValue) => accum+currentValue, 0);
-
+    return R.sum(numbers);
   }
 }
