--- basic/basic-string-repeatN.js
@@ -1,6 +1,0 @@
-class BasicString{
-  repeatWithN(str, count){
-    return [...Array(count)].map(e => e = str).join("");
-}
-}
-module.exports = BasicString;
