+++ basic/strings/basic-string-repeatN.js
@@ -1,0 +1,7 @@
+const R = require('Ramda');
+class BasicString{
+  repeatWithN(str, count){ 
+    return R.join('',R.repeat(str,count))
+  }
+}
+module.exports = BasicString;
