+++ basic/strings/basic-string.js
@@ -1,6 +1,7 @@
+const R = require('Ramda');
 class BasicString{
     repeat(str){
-      return [...Array(5)].map(e => e = str).join("");
+      return R.join('', R.repeat(str, 5))
     }
 }
